+++ extension-compose/src/androidTest/java/com/mapbox/maps/extension/compose/annotation/ViewAnnotationTest.kt
@@ -3,7 +3,6 @@ package com.mapbox.maps.extension.compose.annotation
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.hasText
@@ -15,8 +14,6 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTouchInput
 import androidx.compose.ui.test.swipeLeft
 import com.mapbox.geojson.Point
-import com.mapbox.maps.MapIdle
-import com.mapbox.maps.MapIdleCallback
 import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.ViewAnnotationAnchor
 import com.mapbox.maps.extension.compose.MapboxMap
@@ -27,12 +24,8 @@ import com.mapbox.maps.extension.compose.rememberMapState
 import com.mapbox.maps.viewannotation.annotationAnchor
 import com.mapbox.maps.viewannotation.geometry
 import com.mapbox.maps.viewannotation.viewAnnotationOptions
-import kotlinx.coroutines.launch
-import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.TimeUnit
 
 public class ViewAnnotationTest {
 
@@ -63,24 +56,13 @@ public class ViewAnnotationTest {
 
   @Test
   public fun testSwipeAction() {
-    val mapLoadLatch = CountDownLatch(1)
-    val swipeLatch = CountDownLatch(1)
-    var idleCounter = 0
-    val idleCallback: (MapIdle) -> Unit = {
-      when (idleCounter++) {
-        0 -> mapLoadLatch.countDown()
-        1 -> swipeLatch.countDown()
-      }
-    }
     setMapContent(
       cameraCenter = HELSINKI,
       annotationCenter = HELSINKI,
       visible = true,
-      idleCallback = idleCallback
     ) {
       Text(VIEW_ANNOTATION_TEXT)
     }
-    Assert.assertTrue(mapLoadLatch.await(5, TimeUnit.SECONDS))
     composeTestRule.waitUntil(timeoutMillis = VIEW_APPEAR_TIMEOUT_MS, condition = {
       composeTestRule
         .onAllNodesWithText(VIEW_ANNOTATION_TEXT)
@@ -89,8 +71,11 @@ public class ViewAnnotationTest {
     composeTestRule.onNodeWithTag(MAP_TEST_TAG).performTouchInput {
       swipeLeft(durationMillis = 10L)
     }
-    Assert.assertTrue(swipeLatch.await(3, TimeUnit.SECONDS))
-    composeTestRule.onNodeWithText(VIEW_ANNOTATION_TEXT).assertDoesNotExist()
+    composeTestRule.waitUntil(timeoutMillis = VIEW_DISAPPEAR_TIMEOUT_MS) {
+      composeTestRule
+        .onAllNodesWithText(VIEW_ANNOTATION_TEXT)
+        .fetchSemanticsNodes().isEmpty()
+    }
   }
 
   @Test
@@ -114,11 +99,9 @@ public class ViewAnnotationTest {
     cameraCenter: Point,
     annotationCenter: Point,
     visible: Boolean = true,
-    idleCallback: MapIdleCallback? = null,
     viewAnnotationBlock: @Composable () -> Unit,
   ) {
     composeTestRule.setContent {
-      val coroutineScope = rememberCoroutineScope()
       MapboxMap(
         Modifier
           .fillMaxSize()
@@ -129,15 +112,7 @@ public class ViewAnnotationTest {
             center(cameraCenter)
           }
         },
-        mapState = rememberMapState {
-          coroutineScope.launch {
-            idleCallback?.let { callback ->
-              mapIdleEvents.collect {
-                callback.run(it)
-              }
-            }
-          }
-        },
+        mapState = rememberMapState(),
       ) {
         ViewAnnotation(
           options = viewAnnotationOptions {
@@ -165,5 +140,6 @@ public class ViewAnnotationTest {
     private const val VIEW_ANNOTATION_TEXT = "Test annotation"
     private const val MAP_TEST_TAG = "map_tag"
     private const val VIEW_APPEAR_TIMEOUT_MS = 5000L
+    private const val VIEW_DISAPPEAR_TIMEOUT_MS = 5000L
   }
 }